modules/producer-locations-grpc/app/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
kafka-python==2.0.2
